Bugs fixed.
