+++ src/main/java/edu/hw7/task3/PersonDatabase.java
@@ -1,0 +1,16 @@
+package edu.hw7.task3;
+
+import edu.hw7.Task3;
+import java.util.List;
+
+interface PersonDatabase {
+    void add(Person person);
+
+    void delete(int id);
+
+    List<Person> findByName(String name);
+
+    List<Person> findByAddress(String address);
+
+    List<Person> findByPhone(String phone);
+}
