+++ src/test/java/edu/hw7/Task3dot5Test.java
@@ -1,6 +1,8 @@
 package edu.hw7;
 
 import java.util.List;
+import edu.hw7.task3.Person;
+import edu.hw7.task3.PersonDatabaseReadWriteLockRealization;
 import org.junit.jupiter.api.Test;
 import static java.lang.Thread.sleep;
 import static org.assertj.core.api.AssertionsForClassTypes.assertThat;
@@ -8,21 +10,21 @@ import static org.assertj.core.api.AssertionsForClassTypes.assertThat;
 public class Task3dot5Test {
     @Test
     void givenPersonDBWhenAddingAndFindingByNamePeopleThenReturn() throws InterruptedException {
-        Task3.PersonDatabaseReadWriteLockRealization
-            personDatabaseReadWriteLockRealization = new Task3.PersonDatabaseReadWriteLockRealization();
+        PersonDatabaseReadWriteLockRealization
+            personDatabaseReadWriteLockRealization = new PersonDatabaseReadWriteLockRealization();
 
         Runnable runnable1 = new Runnable() {
             @Override
             public void run() {
-                personDatabaseReadWriteLockRealization.add(new Task3.Person(1, "Tim", "Gagarina 30", "+7478447"));
-                personDatabaseReadWriteLockRealization.add(new Task3.Person(2, "Tom", "Torina 10", "+74121257"));
+                personDatabaseReadWriteLockRealization.add(new Person(1, "Tim", "Gagarina 30", "+7478447"));
+                personDatabaseReadWriteLockRealization.add(new Person(2, "Tom", "Torina 10", "+74121257"));
             }
         },
             runnable2 = new Runnable() {
                 @Override
                 public void run() {
-                    personDatabaseReadWriteLockRealization.add(new Task3.Person(5, "Kamila", "Bimireno 30", "+2352351"));
-                    personDatabaseReadWriteLockRealization.add(new Task3.Person(
+                    personDatabaseReadWriteLockRealization.add(new Person(5, "Kamila", "Bimireno 30", "+2352351"));
+                    personDatabaseReadWriteLockRealization.add(new Person(
                         12,
                         "Niemand",
                         "Kokorina 10",
@@ -36,30 +38,31 @@ public class Task3dot5Test {
         th1.start();
         th2.start();
 
-        sleep(100);
+        th1.join();
+        th2.join();
 
         assertThat(personDatabaseReadWriteLockRealization.findByName("Tim")).isEqualTo(
-            List.of(new Task3.Person(1, "Tim", "Gagarina 30", "+7478447")
+            List.of(new Person(1, "Tim", "Gagarina 30", "+7478447")
             ));
     }
 
     @Test
     void givenPersonDBWhenAddingAndFindingByAddressPeopleThenReturn() throws InterruptedException {
-        Task3.PersonDatabaseReadWriteLockRealization
-            personDatabaseReadWriteLockRealization = new Task3.PersonDatabaseReadWriteLockRealization();
+        PersonDatabaseReadWriteLockRealization
+            personDatabaseReadWriteLockRealization = new PersonDatabaseReadWriteLockRealization();
 
         Runnable runnable1 = new Runnable() {
             @Override
             public void run() {
-                personDatabaseReadWriteLockRealization.add(new Task3.Person(1, "Tim", "Gagarina 30", "+7478447"));
-                personDatabaseReadWriteLockRealization.add(new Task3.Person(2, "Tom", "Torina 10", "+74121257"));
+                personDatabaseReadWriteLockRealization.add(new Person(1, "Tim", "Gagarina 30", "+7478447"));
+                personDatabaseReadWriteLockRealization.add(new Person(2, "Tom", "Torina 10", "+74121257"));
             }
         },
             runnable2 = new Runnable() {
                 @Override
                 public void run() {
-                    personDatabaseReadWriteLockRealization.add(new Task3.Person(5, "Kamila", "Bimireno 30", "+2352351"));
-                    personDatabaseReadWriteLockRealization.add(new Task3.Person(
+                    personDatabaseReadWriteLockRealization.add(new Person(5, "Kamila", "Bimireno 30", "+2352351"));
+                    personDatabaseReadWriteLockRealization.add(new Person(
                         12,
                         "Niemand",
                         "Kokorina 10",
@@ -73,30 +76,31 @@ public class Task3dot5Test {
         th1.start();
         th2.start();
 
-        sleep(100);
+        th1.join();
+        th2.join();
 
         assertThat(personDatabaseReadWriteLockRealization.findByAddress("Torina 10")).isEqualTo(
-            List.of(new Task3.Person(2, "Tom", "Torina 10", "+74121257")
+            List.of(new Person(2, "Tom", "Torina 10", "+74121257")
             ));
     }
 
     @Test
     void givenPersonDBWhenAddingAndFindingByPhoneNumberPeopleThenReturn() throws InterruptedException {
-        Task3.PersonDatabaseReadWriteLockRealization
-            personDatabaseReadWriteLockRealization = new Task3.PersonDatabaseReadWriteLockRealization();
+        PersonDatabaseReadWriteLockRealization
+            personDatabaseReadWriteLockRealization = new PersonDatabaseReadWriteLockRealization();
 
         Runnable runnable1 = new Runnable() {
             @Override
             public void run() {
-                personDatabaseReadWriteLockRealization.add(new Task3.Person(1, "Tim", "Gagarina 30", "+7478447"));
-                personDatabaseReadWriteLockRealization.add(new Task3.Person(2, "Tom", "Torina 10", "+74121257"));
+                personDatabaseReadWriteLockRealization.add(new Person(1, "Tim", "Gagarina 30", "+7478447"));
+                personDatabaseReadWriteLockRealization.add(new Person(2, "Tom", "Torina 10", "+74121257"));
             }
         },
             runnable2 = new Runnable() {
                 @Override
                 public void run() {
-                    personDatabaseReadWriteLockRealization.add(new Task3.Person(5, "Kamila", "Bimireno 30", "+2352351"));
-                    personDatabaseReadWriteLockRealization.add(new Task3.Person(
+                    personDatabaseReadWriteLockRealization.add(new Person(5, "Kamila", "Bimireno 30", "+2352351"));
+                    personDatabaseReadWriteLockRealization.add(new Person(
                         12,
                         "Niemand",
                         "Kokorina 10",
@@ -110,33 +114,34 @@ public class Task3dot5Test {
         th1.start();
         th2.start();
 
-        sleep(100);
+        th1.join();
+        th2.join();
 
         assertThat(personDatabaseReadWriteLockRealization.findByPhone("+2352351")).isEqualTo(
-            List.of(new Task3.Person(5, "Kamila", "Bimireno 30", "+2352351")
+            List.of(new Person(5, "Kamila", "Bimireno 30", "+2352351")
             ));
     }
 
     @Test
     void givenPersonDBWhenParallelAddingAndFindingPeopleByNameThenReturn() throws InterruptedException {
-        Task3.PersonDatabaseReadWriteLockRealization
-            personDatabaseReadWriteLockRealization = new Task3.PersonDatabaseReadWriteLockRealization();
+        PersonDatabaseReadWriteLockRealization
+            personDatabaseReadWriteLockRealization = new PersonDatabaseReadWriteLockRealization();
 
         Runnable runnable1 = new Runnable() {
             @Override
             public void run() {
-                personDatabaseReadWriteLockRealization.add(new Task3.Person(1, "Tim", "Gagarina 30", "+7478447"));
-                personDatabaseReadWriteLockRealization.add(new Task3.Person(2, "Tom", "Torina 10", "+74121257"));
+                personDatabaseReadWriteLockRealization.add(new Person(1, "Tim", "Gagarina 30", "+7478447"));
+                personDatabaseReadWriteLockRealization.add(new Person(2, "Tom", "Torina 10", "+74121257"));
 
                 assertThat(personDatabaseReadWriteLockRealization.findByName("Tom"))
-                    .isEqualTo(new Task3.Person(2, "Tom", "Torina 10", "+74121257"));
+                    .isEqualTo(new Person(2, "Tom", "Torina 10", "+74121257"));
             }
         },
             runnable2 = new Runnable() {
                 @Override
                 public void run() {
-                    personDatabaseReadWriteLockRealization.add(new Task3.Person(5, "Kamila", "Bimireno 30", "+2352351"));
-                    personDatabaseReadWriteLockRealization.add(new Task3.Person(
+                    personDatabaseReadWriteLockRealization.add(new Person(5, "Kamila", "Bimireno 30", "+2352351"));
+                    personDatabaseReadWriteLockRealization.add(new Person(
                         12,
                         "Niemand",
                         "Kokorina 10",
@@ -153,28 +158,28 @@ public class Task3dot5Test {
 
     @Test
     void givenPersonDBWhenParallelAddingAndFindingPeopleByAddressThenReturn() throws InterruptedException {
-        Task3.PersonDatabaseReadWriteLockRealization
-            personDatabaseReadWriteLockRealization = new Task3.PersonDatabaseReadWriteLockRealization();
+        PersonDatabaseReadWriteLockRealization
+            personDatabaseReadWriteLockRealization = new PersonDatabaseReadWriteLockRealization();
 
         Runnable runnable1 = new Runnable() {
             @Override
             public void run() {
-                personDatabaseReadWriteLockRealization.add(new Task3.Person(1, "Tim", "Gagarina 30", "+7478447"));
-                personDatabaseReadWriteLockRealization.add(new Task3.Person(2, "Tom", "Torina 10", "+74121257"));
+                personDatabaseReadWriteLockRealization.add(new Person(1, "Tim", "Gagarina 30", "+7478447"));
+                personDatabaseReadWriteLockRealization.add(new Person(2, "Tom", "Torina 10", "+74121257"));
             }
         },
             runnable2 = new Runnable() {
                 @Override
                 public void run() {
-                    personDatabaseReadWriteLockRealization.add(new Task3.Person(5, "Kamila", "Bimireno 30", "+2352351"));
-                    personDatabaseReadWriteLockRealization.add(new Task3.Person(
+                    personDatabaseReadWriteLockRealization.add(new Person(5, "Kamila", "Bimireno 30", "+2352351"));
+                    personDatabaseReadWriteLockRealization.add(new Person(
                         12,
                         "Niemand",
                         "Kokorina 10",
                         "+72352357"
                     ));
                     assertThat(personDatabaseReadWriteLockRealization.findByAddress("Bimireno 30"))
-                        .isEqualTo(new Task3.Person(5, "Kamila", "Bimireno 30", "+2352351"));
+                        .isEqualTo(new Person(5, "Kamila", "Bimireno 30", "+2352351"));
                 }
             };
         Thread th1 = new Thread(runnable1);
@@ -186,28 +191,28 @@ public class Task3dot5Test {
 
     @Test
     void givenPersonDBWhenParallelAddingAndFindingPeopleByPhoneThenReturn() throws InterruptedException {
-        Task3.PersonDatabaseReadWriteLockRealization
-            personDatabaseReadWriteLockRealization = new Task3.PersonDatabaseReadWriteLockRealization();
+        PersonDatabaseReadWriteLockRealization
+            personDatabaseReadWriteLockRealization = new PersonDatabaseReadWriteLockRealization();
 
         Runnable runnable1 = new Runnable() {
             @Override
             public void run() {
-                personDatabaseReadWriteLockRealization.add(new Task3.Person(1, "Tim", "Gagarina 30", "+7478447"));
-                personDatabaseReadWriteLockRealization.add(new Task3.Person(2, "Tom", "Torina 10", "+74121257"));
+                personDatabaseReadWriteLockRealization.add(new Person(1, "Tim", "Gagarina 30", "+7478447"));
+                personDatabaseReadWriteLockRealization.add(new Person(2, "Tom", "Torina 10", "+74121257"));
             }
         },
             runnable2 = new Runnable() {
                 @Override
                 public void run() {
-                    personDatabaseReadWriteLockRealization.add(new Task3.Person(5, "Kamila", "Bimireno 30", "+2352351"));
-                    personDatabaseReadWriteLockRealization.add(new Task3.Person(
+                    personDatabaseReadWriteLockRealization.add(new Person(5, "Kamila", "Bimireno 30", "+2352351"));
+                    personDatabaseReadWriteLockRealization.add(new Person(
                         12,
                         "Niemand",
                         "Kokorina 10",
                         "+72352357"
                     ));
                     assertThat(personDatabaseReadWriteLockRealization.findByPhone("+72352357"))
-                        .isEqualTo(List.of(new Task3.Person(
+                        .isEqualTo(List.of(new Person(
                             12,
                             "Niemand",
                             "Kokorina 10",
